Add CMake project and fix clang warnings (#139)

+++ DirectXTex/DirectXTexCompress.cpp
@@ -9,7 +9,7 @@
 // http://go.microsoft.com/fwlink/?LinkId=248926
 //-------------------------------------------------------------------------------------
 
-#include "DirectXTexp.h"
+#include "DirectXTexP.h"
 
 #ifdef _OPENMP
 #include <omp.h>
@@ -120,25 +120,25 @@ namespace
 
                 ptrdiff_t bytesLeft = pEnd - sptr;
                 assert(bytesLeft > 0);
-                size_t bytesToRead = std::min<size_t>(rowPitch, bytesLeft);
+                size_t bytesToRead = std::min<size_t>(rowPitch, static_cast<size_t>(bytesLeft));
                 if (!_LoadScanline(&temp[0], pw, sptr, bytesToRead, format))
                     return E_FAIL;
 
                 if (ph > 1)
                 {
-                    bytesToRead = std::min<size_t>(rowPitch, bytesLeft - rowPitch);
+                    bytesToRead = std::min<size_t>(rowPitch, static_cast<size_t>(bytesLeft) - rowPitch);
                     if (!_LoadScanline(&temp[4], pw, sptr + rowPitch, bytesToRead, format))
                         return E_FAIL;
 
                     if (ph > 2)
                     {
-                        bytesToRead = std::min<size_t>(rowPitch, bytesLeft - rowPitch * 2);
+                        bytesToRead = std::min<size_t>(rowPitch, static_cast<size_t>(bytesLeft) - rowPitch * 2);
                         if (!_LoadScanline(&temp[8], pw, sptr + rowPitch * 2, bytesToRead, format))
                             return E_FAIL;
 
                         if (ph > 3)
                         {
-                            bytesToRead = std::min<size_t>(rowPitch, bytesLeft - rowPitch * 3);
+                            bytesToRead = std::min<size_t>(rowPitch, static_cast<size_t>(bytesLeft) - rowPitch * 3);
                             if (!_LoadScanline(&temp[12], pw, sptr + rowPitch * 3, bytesToRead, format))
                                 return E_FAIL;
                         }
@@ -260,7 +260,7 @@ namespace
 
             ptrdiff_t bytesLeft = pEnd - pSrc;
             assert(bytesLeft > 0);
-            size_t bytesToRead = std::min<size_t>(rowPitch, bytesLeft);
+            size_t bytesToRead = std::min<size_t>(rowPitch, size_t(bytesLeft));
 
             __declspec(align(16)) XMVECTOR temp[16];
             if (!_LoadScanline(&temp[0], pw, pSrc, bytesToRead, format))
@@ -268,19 +268,19 @@ namespace
 
             if (ph > 1)
             {
-                bytesToRead = std::min<size_t>(rowPitch, bytesLeft - rowPitch);
+                bytesToRead = std::min<size_t>(rowPitch, size_t(bytesLeft - rowPitch));
                 if (!_LoadScanline(&temp[4], pw, pSrc + rowPitch, bytesToRead, format))
                     fail = true;
 
                 if (ph > 2)
                 {
-                    bytesToRead = std::min<size_t>(rowPitch, bytesLeft - rowPitch * 2);
+                    bytesToRead = std::min<size_t>(rowPitch, size_t(bytesLeft - rowPitch * 2));
                     if (!_LoadScanline(&temp[8], pw, pSrc + rowPitch * 2, bytesToRead, format))
                         fail = true;
 
                     if (ph > 3)
                     {
-                        bytesToRead = std::min<size_t>(rowPitch, bytesLeft - rowPitch * 3);
+                        bytesToRead = std::min<size_t>(rowPitch, size_t(bytesLeft - rowPitch * 3));
                         if (!_LoadScanline(&temp[12], pw, pSrc + rowPitch * 3, bytesToRead, format))
                             fail = true;
                     }
